fix: Cloudformation allow changeset updates
This commit fixes two issues with Cloudformation change set updates.

Firstly, `moto` returns an exception: `KeyError: 'parameter_value'` for
existing parameters if the change set is created without the
`UsePreviousTemplate` argument.

This was fixed by removing the condition for `UsePreviousTemplate` in
the parameter list generation.

The second issue was default values defined in the Parameter section in
the Cloudformation template are not included in the parameter arguments
to `CloudFormationBackend.create_change_set` when the change set is
being parsed/validated.

This results in a change set creation failing due to missing
Parameter values.

The second issue was fixed by replacing references to
`FakeStack.parameters` with `FakeStack.stack_parameters` which returns
the resource parameters vs just the parameters provided to
Cloudformation create/update stack operations.