Fix mypy
We have to essentially make params a magic object (Any) because otherwise we'd have a type error
every time we use .get (which could return None).  The alternative is to literally generate statically
typed Python classes for every Input shape in the Botocore models, but... it just doesn't seem worth
it (at the moment, anyway).  We could also do something where the response methods specify which parameter
names they want (with type annotations) and then we supply them dynamically, but that would be a ton of
work.  Still mulling this one over, but I think we need to be fully dynamic for the moment to ease integration.  Also, keep in my mind that all of this currently works fine at runtime.