Core: refactor ServiceException serialization
The ServiceException handler has been refactored similar to the HTTPException handler,
and no longer maintains a reference to `e` outside the Except block.

It may have always been a bad idea to maintain an exception reference that persists
outside the Except block, but it actually causes a very strange issue when Moto is
in proxy mode.  After an error is raised and then serialized as a response, the very
next Boto3 call to the proxy will result in a read timeout (after the default 60 seconds).