Fix `get_value` fallback for dicts
There's an edge-case when we're trying to get the key `values` from a dict.
If the key exists, no problem; if it doesn't, we default to `getattr` and
wind up with the standard dict `.values()` callable, which is clearly not
what we want.

Note: with this change, we can't fully support getting values from dicts
that also contain properties.  It's not clear if any structures like that
exist in Moto, but they would have to be refactored if encountered going
forward.