Fix remaining tests
The ec2 model makes a distinction between a NetworkInterface object
and an InstanceNetworkInterface object.  Other attributes follow the
same naming convention, so I was able to use the serializer's built-in
alias list to set properties for the instance-specific ENI stuff.

Might want to have a wrapper class in the Instances model instead so
I don't have to touch NetworkInterfaces at all...