Add MAX_FORM_MEMORY_SIZE setting
It's possible with some of the XML-based protocols to exceed the Werkzeug default,
which results in a 413 error that isn't parseable by the Boto3 client.  Some backends,
like SQS, have dedicated errors for payloads/parameters being over a certain max size,
so we need to make sure the request doesn't get rejected at the lower HTTP server level.