fix: Optimize request parsing.
When sending a large `Body` to `sagemaker_runtime.invoke_endpoint` the
response times would just explode. The regular expression used for
sorting the keys had some bad backtracking and performance behavior.

To improve the situation, we can simplify the regex and make use of the
Match object returned from the search to do the splitting. This
implementation is intended to retain the existing behavior but if the
`member` suffix in the first element of the tuple isn't needed then the
magic `+7` could be dropped.