Ensure default response is equivalent
There's some obviously wrong behavior here, but I'm erroring on the side of making
this as close as possible to the original XML-templated response.

Only one attribute assertion had to change: DeleteOnTermination for the NetworkInterface.Attachment
I considered changing it, but wasn't sure if there's some behavior associated with that value, so
I opted to go with what the model (has always) said (because True was hardcoded in the XML template).