Initial commit
I decided to make everything bytes in the model because it's more explict than using string,
because you don't know if that string is a raw string or a base64-encoded string.

I think it's the right thing to do, but I don't like how many files needed to be touched.
Also, the Botocore models all use string and I don't really want to override that many
shapes to have automatically base64 encoded/decoded.

Still considering if there's another way to get the explicitness of a unique type in the
model without having to do this much extra work and override so many Botocore shapes...